Move ToggleableTimer to a new file.
